Decouple disklayout and add var fstype
Make a separate disklayouts directory in templates
where the disklayouts can be stored with different names.

Currently there are
- server (root and swap fs)
- minimal (root, home and swap fs)
- pcidss (standard pcidss fs layout)

Futher the variable fstype has been added to change the
filesystem type for all installations or just for one OS.

Set the disk size to 20G which at this point only works with
the default or server disk layout for others the size has to
be adjusted.

Closes #3
Closes #4